fix: theme switcher hydration error
