--- src/app/(main)/layout.tsx
@@ -1,9 +1,0 @@
-"use client";
-
-export default function Layout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
-  return <>{children}</>;
-}
